backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.10.6
python-dotenv==1.0.1
